+++ tests/testthat/test-fst_ld.R
@@ -242,18 +242,28 @@ test_that("fis bootstrapping",{
   expect_true(all(bs1_res$weighted.means[bs1_res$weighted.means$facet == "pop",]$snp.subfacet == ".base"))
 })
 
-# test_that("global fst",{
-#   tdfst <- calc_pairwise_fst(.internal.data$test_snps, "pop", "wc", global = TRUE)
-#   tdfst <- get.snpR.stats(tdfst, "pop", "fst")
-#   expect_equal(round(tdfst$pairwise$fst, 4), 
-#                round(c(0.034091, 0, -0.122941, 0, -0.09375, -0.026012, 
-#                        0.166667, 0.107143, -0.057692, -0.086957, 0.016548), 
-#                      4)) # values from pegas, also double checked by hand. Note that heirfstat slightly disagrees on a few of these (where the allele is fixed in one loci)
-#   expect_equal(round(tdfst$weighted.means$weighted_mean_fst, 5), -0.00343) # hand calced due to weighting and ratio of averages approach
-#   
-#   # var comps
-#   expect_false(any(paste0("var_comp_", c("a", "b", "c")) %in% colnames(tdfst$pairwise))) # didn't ask for var comps
-#   tdfst <- calc_pairwise_fst(.internal.data$test_snps, "pop", "wc", keep_components = TRUE)
-#   tdfst <- get.snpR.stats(tdfst, "pop", "fst")
-#   expect_true(all(paste0("var_comp_", c("a", "b", "c")) %in% colnames(tdfst$pairwise))) # did ask for var comps
-# })
+test_that("global fst",{
+  
+  # should yield the same as global = FALSE for two comps
+  tdfst <- calc_global_fst(.internal.data$test_snps, "pop")
+  tdfst <- get.snpR.stats(tdfst, "pop", "fst")
+  expect_equal(round(tdfst$pairwise$fst, 4),
+               round(c(0.034091, 0, -0.122941, 0, -0.09375, -0.026012,
+                       0.166667, 0.107143, -0.057692, -0.086957, 0.016548),
+                     4)) # values from pegas, also double checked by hand. Note that heirfstat slightly disagrees on a few of these (where the allele is fixed in one loci)
+  expect_equal(round(tdfst$weighted.means$weighted_mean_fst, 5), -0.00343) # hand calced due to weighting and ratio of averages approach
+  
+  # actual multipop
+  tdfst <- calc_global_fst(stickSNPs, "pop")
+  tdfst <- get.snpR.stats(tdfst, "pop", "fst")
+  
+  # peg <- snpR::format_snps(stickSNPs, facets = "pop", output = "adegenet")
+  # peg <- pegas::genind2loci(peg)
+  # peg_fst <- pegas::Fst(peg)
+  # round(peg_fst[,2], 3)
+  
+  expect_equivalent(round(tdfst$pairwise$fst, 3)[1:10],
+                    c(0.078, -0.002, -0.025, 0.143, 0.011, 0.136, 0.059, 0.08, 0.04, 0.025)) # from pegas, see above
+  
+  expect_equal(round(tdfst$weighted.means$weighted_mean_fst, 3), 0.094)
+})